Add files via upload
This is a script I wrote in the latest version of RStudio to parse these JSON files: https://mtgjson.com. I coerce the JSON into a data-frame and make an animated plot comparing CMC-to-Rules-Text-Length for each color of card.

To run it on your own machine, and use the data-frame for your own projects, you'll need to change some lines. My comments should help you understand how the inconsistent JSON list is parsed into a data-frame.